requirements.txt
torch>=2.3.0
transformers>=4.51.0
qwen-vl-utils>=0.0.10
accelerate>=0.30.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
requests>=2.31.0          # required for predict_vllm.py
vllm>=0.8.0               # required only for serving; install separately on the GPU server
